refactor(wasm): ship the AST typings with the wasm package
A standalone @gengjiawen/monkey-ast-types package was one more build to
order and one more publish to sequence, for declarations that only
describe what the wasm parser emits. Ship them with the parser instead:
wasm/src/ast_types.d.ts is appended to the generated monkey_wasm.d.ts
through wasm-bindgen's typescript_custom_section, so the npm package
that carries the parser also carries the types for its output and the
two cannot drift apart.

The linter, minifier and prettier plugin now re-export the nodes from
@gengjiawen/monkey-wasm and the type checker imports them directly. The
small runtime accessors (identifierName, setIdentifierName, tokenType)
move into the packages that call them; isTypeAnnotation and
printTypeAnnotation had no callers and are gone. The package itself is
deleted along with its CI step, publish step, Vercel pre-build and
version-sync entries.
