+++ .github/workflows/release-please.yml
@@ -85,15 +85,6 @@ jobs:
         env:
           NODE_AUTH_TOKEN: ${{secrets.NPM_TOKEN}}
         if: ${{ steps.release.outputs.release_created }}
-      # The shared AST types publish first: the minifier, linter, prettier
-      # plugin and type checker all declare a dependency on them.
-      - run: pnpm install --filter @gengjiawen/monkey-ast-types... --frozen-lockfile && pnpm --filter @gengjiawen/monkey-ast-types run build
-        if: ${{ steps.release.outputs.release_created }}
-      - run: npm publish --access public
-        working-directory: packages/monkey-ast-types
-        env:
-          NODE_AUTH_TOKEN: ${{secrets.NPM_TOKEN}}
-        if: ${{ steps.release.outputs.release_created }}
       # Ahead of the VS Code extension, which bundles the checker into its
       # diagnostics and so needs the built declarations.
       - run: pnpm install --filter @gengjiawen/monkey-typechecker... --frozen-lockfile && pnpm --filter @gengjiawen/monkey-typechecker run build
