+++ .github/workflows/rust.yml
@@ -39,13 +39,6 @@ jobs:
         run: cd wasm && wasm-pack build --release --scope=gengjiawen
       - name: Run wasm tests
         run: cd wasm && wasm-pack test --node
-      # Every JS package below reads the AST through this one, and they import
-      # it by name, so its declarations have to exist before they build.
-      - name: Build and test shared AST types
-        run: |
-          pnpm install --filter @gengjiawen/monkey-ast-types --frozen-lockfile
-          pnpm --filter @gengjiawen/monkey-ast-types build
-          pnpm --filter @gengjiawen/monkey-ast-types test
       - name: Build and test minifier
         run: |
           pnpm install --filter @gengjiawen/monkey-minifier... --frozen-lockfile
