--- packages/monkey-ast-types/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "@gengjiawen/monkey-ast-types",
-  "version": "1.2.0",
-  "description": "Shared TypeScript definitions for the Monkey AST JSON emitted by @gengjiawen/monkey-wasm",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "scripts": {
-    "build": "tsc",
-    "test": "vitest run",
-    "dev": "tsc --watch",
-    "prepublishOnly": "npm run build"
-  },
-  "keywords": [
-    "monkey",
-    "ast",
-    "types"
-  ],
-  "author": "gengjiawen <technicalcute@gmail.com>",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/gengjiawen/monkey-rust"
-  },
-  "devDependencies": {
-    "@types/node": "^24.0.0",
-    "typescript": "^7.0.2",
-    "vitest": "^4.1.10"
-  },
-  "files": [
-    "dist",
-    "README.md"
-  ]
-}
