+++ packages/monkey-linter/package.json
@@ -32,7 +32,6 @@
     "url": "https://github.com/gengjiawen/monkey-rust"
   },
   "dependencies": {
-    "@gengjiawen/monkey-ast-types": "^1.2.0",
     "@gengjiawen/monkey-wasm": "^1.2.0"
   },
   "devDependencies": {
