+++ packages/monkey-minifier/src/types.ts
@@ -1,13 +1,20 @@
-// AST node definitions live in `@gengjiawen/monkey-ast-types`, shared with the
-// linter, the prettier plugin and the type checker. Re-exported here so the
-// rest of the minifier keeps importing from one place. The helpers are listed
-// by name so bundlers see a static re-export instead of having to resolve the
-// CommonJS star at runtime.
-export type * from '@gengjiawen/monkey-ast-types'
-export {
-  identifierName,
-  isTypeAnnotation,
-  printTypeAnnotation,
-  setIdentifierName,
-  tokenType,
-} from '@gengjiawen/monkey-ast-types'
+// The AST node definitions ship inside `@gengjiawen/monkey-wasm` itself:
+// wasm/src/ast_types.d.ts is appended to the generated declarations, so the
+// tree described there is always the one the linked parser build emits.
+// Re-exported here so the rest of the minifier keeps importing from one place;
+// the accessors below are the only runtime code needed on top.
+export type * from '@gengjiawen/monkey-wasm'
+
+import type { LetStatement, Token } from '@gengjiawen/monkey-wasm'
+
+export function identifierName(statement: LetStatement): string {
+  return statement.identifier.name
+}
+
+export function setIdentifierName(statement: LetStatement, name: string): void {
+  statement.identifier.name = name
+}
+
+export function tokenType(token: Token): string {
+  return token.kind.type
+}
