+++ packages/prettier-plugin-monkey/package.json
@@ -27,7 +27,6 @@
     "prettier": "^3.0.0"
   },
   "dependencies": {
-    "@gengjiawen/monkey-ast-types": "^1.2.0",
     "@gengjiawen/monkey-wasm": "^1.2.0"
   },
   "devDependencies": {
