+++ scripts/bump_cargo_packages.ts
@@ -124,24 +124,6 @@ try {
   console.warn('Failed to update playground dependency:', e)
 }
 
-// Also keep monkey-ast-types in sync. It has no dependencies of its own; the
-// minifier, linter, prettier plugin and type checker all depend on it, so it is
-// bumped first.
-try {
-  const astTypesPkgPath = repoPath(
-    'packages',
-    'monkey-ast-types',
-    'package.json'
-  )
-  const astTypes = readPackageJson(astTypesPkgPath)
-
-  if (syncPackageVersion(astTypes, 'monkey-ast-types', nextVersion)) {
-    writePackageJson(astTypesPkgPath, astTypes)
-  }
-} catch (e) {
-  console.warn('Failed to update monkey-ast-types version:', e)
-}
-
 // Also keep prettier-plugin-monkey package version and wasm dependency in sync
 try {
   const prettierPluginPkgPath = repoPath(
@@ -160,13 +142,6 @@ try {
       prettierPlugin,
       'prettier-plugin-monkey',
       '@gengjiawen/monkey-wasm',
-      `^${nextVersion}`
-    ) || prettierPluginChanged
-  prettierPluginChanged =
-    syncDependencyRange(
-      prettierPlugin,
-      'prettier-plugin-monkey',
-      '@gengjiawen/monkey-ast-types',
       `^${nextVersion}`
     ) || prettierPluginChanged
 
@@ -200,13 +175,6 @@ try {
       '@gengjiawen/monkey-wasm',
       `^${nextVersion}`
     ) || minifierChanged
-  minifierChanged =
-    syncDependencyRange(
-      minifier,
-      'monkey-minifier',
-      '@gengjiawen/monkey-ast-types',
-      `^${nextVersion}`
-    ) || minifierChanged
 
   if (minifierChanged) {
     writePackageJson(minifierPkgPath, minifier)
@@ -229,13 +197,6 @@ try {
       linter,
       'monkey-lint',
       '@gengjiawen/monkey-wasm',
-      `^${nextVersion}`
-    ) || linterChanged
-  linterChanged =
-    syncDependencyRange(
-      linter,
-      'monkey-lint',
-      '@gengjiawen/monkey-ast-types',
       `^${nextVersion}`
     ) || linterChanged
 
@@ -265,13 +226,6 @@ try {
       typechecker,
       'monkey-typechecker',
       '@gengjiawen/monkey-wasm',
-      `^${nextVersion}`
-    ) || typecheckerChanged
-  typecheckerChanged =
-    syncDependencyRange(
-      typechecker,
-      'monkey-typechecker',
-      '@gengjiawen/monkey-ast-types',
       `^${nextVersion}`
     ) || typecheckerChanged
